Move mEstimator into separate file
